requirements.txt
flask
redis
gunicorn
gevent
cassandra-driver
